Upgrade to comply with new use of MealRepo
