+++ app/src/androidTest/java/com/github/se/polyfit/ui/components/nutrition/NutritionalInformationTest.kt
@@ -10,7 +10,6 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performScrollToIndex
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.se.polyfit.model.data.User
 import com.github.se.polyfit.model.meal.Meal
 import com.github.se.polyfit.model.meal.MealOccasion
 import com.github.se.polyfit.model.nutritionalInformation.MeasurementUnit
@@ -89,9 +88,7 @@ class NutritionalInformationTest : TestCase() {
 
   private fun setup(meal: Meal) {
     composeTestRule.setContent {
-      NutritionalInformation(
-          mealViewModel =
-              MealViewModel(User(id = "testUserID"), mockk()).apply { setMealData(meal) })
+      NutritionalInformation(mealViewModel = MealViewModel(mockk()).apply { setMealData(meal) })
     }
   }
