+++ app/src/androidTest/java/com/github/se/polyfit/ui/flow/AddMealFlowTest.kt
@@ -5,7 +5,6 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.se.polyfit.model.data.User
 import com.github.se.polyfit.ui.navigation.Navigation
 import com.github.se.polyfit.ui.navigation.Route
 import com.github.se.polyfit.ui.screen.HomeScreen
@@ -37,10 +36,7 @@ class AddMealFlowTest {
         composable(Route.Home) { HomeScreen() }
         composable(Route.AddMeal) {
           AddMealFlow(
-              navigation::goBack,
-              navigation::navigateToHome,
-              "testUserID",
-              MealViewModel(User(id = "testUserID"), mockk()))
+              navigation::goBack, navigation::navigateToHome, "testUserID", MealViewModel(mockk()))
         }
       }
       navigation.navigateToAddMeal()
